Refactoring characteristics: does not compile yet. The CharacteristicBase will have an implementation that works with a buffer as value. The CharacteristicGeneric will be convienence variations of it. This centralizes the implementation, and should make it easier to follow.
